Fixed last few warnings for GCC!

+++ codemp/rd-vanilla/tr_font.cpp
@@ -108,7 +108,7 @@ struct ThaiCodes_t
 
 	int GetWidth( int iGlyphIndex )
 	{
-		if (iGlyphIndex < m_viGlyphWidths.size())
+		if (iGlyphIndex < (int)m_viGlyphWidths.size())
 		{
 			return m_viGlyphWidths[ iGlyphIndex ];
 		}
@@ -1369,7 +1369,7 @@ int RE_Font_StrLenPixels(const char *psText, const int iFontHandle, const float 
 		{
 			int iPixelAdvance = curfont->GetLetterHorizAdvance( uiLetter );
 	
-			float fValue = iPixelAdvance * ((uiLetter > g_iNonScaledCharRange) ? fScaleA : fScale);
+			float fValue = iPixelAdvance * ((uiLetter > (unsigned)g_iNonScaledCharRange) ? fScaleA : fScale);
 			iThisWidth += curfont->mbRoundCalcs ? Round( fValue ) : fValue;
 			if (iThisWidth > iMaxWidth)
 			{
@@ -1583,7 +1583,7 @@ void RE_Font_DrawString(int ox, int oy, const char *psText, const float *rgba, c
 				pLetter = curfont->GetLetter('.');
 			}
 
-			float fThisScale = uiLetter > g_iNonScaledCharRange ? fScaleA : fScale;
+			float fThisScale = uiLetter > (unsigned)g_iNonScaledCharRange ? fScaleA : fScale;
 
 			// sigh, super-language-specific hack...
 			//
@@ -1605,7 +1605,7 @@ void RE_Font_DrawString(int ox, int oy, const char *psText, const float *rgba, c
 				}
 
 				RE_StretchPic ( x + Round(pLetter->horizOffset * fScale), // float x
-								(uiLetter > g_iNonScaledCharRange) ? y - iAsianYAdjust : y,	// float y
+								(uiLetter > (unsigned)g_iNonScaledCharRange) ? y - iAsianYAdjust : y,	// float y
 								curfont->mbRoundCalcs ? Round(pLetter->width * fThisScale) : pLetter->width * fThisScale,	// float w
 								curfont->mbRoundCalcs ? Round(pLetter->height * fThisScale) : pLetter->height * fThisScale, // float h
 								pLetter->s,						// float s1
@@ -1727,11 +1727,11 @@ void R_ReloadFonts_f(void)
 		//
 		// and re-register our fonts in the same order as before (note that some menu items etc cache the string lengths so really a vid_restart is better, but this is just for my testing)
 		//
-		for (int iFont = 0; iFont < vstrFonts.size(); iFont++)
+		for (size_t iFont = 0; iFont < vstrFonts.size(); iFont++)
 		{
 #ifdef _DEBUG
 			int iNewFontHandle = RE_RegisterFont( vstrFonts[iFont].c_str() );
-			assert( iNewFontHandle == iFont+1 );
+			assert( iNewFontHandle == (int)(iFont+1) );
 #else
 			RE_RegisterFont( vstrFonts[iFont].c_str() );
 #endif
